requirements.txt
flask==2.1.2
flask_script==2.0.6
Flask-Migrate==3.1.0
psycopg2==2.9.1
marshmallow==3.14.1
pyjwt==2.4.0
requests==2.27.1
pytest==7.1.2
#tornado==6.1